Added MCF5225X GPIO device I2C now uses GPIO device Minor changes in configuration files
git-svn-id: svn+ssh://svn.code.sf.net/p/ethernut/code/branches/devnut_freescale@4580 1ca430b3-8271-4a6b-9b0e-3dcac98a8924

+++ nut/include/sys/types.h
@@ -249,7 +249,7 @@ typedef void *HANDLE;
 #elif defined(__H8300__) || defined(__H8300H__) || defined(__H8300S__)
     typedef unsigned short ureg_t;
 #elif defined(__m68k__)
-    typedef unsigned short ureg_t;
+    typedef uint_fast8_t ureg_t;
 #else
     typedef unsigned short ureg_t;
 #endif
@@ -270,7 +270,7 @@ typedef void *HANDLE;
 #elif defined(__H8300__) || defined(__H8300H__) || defined(__H8300S__)
     typedef unsigned short reg_t;
 #elif defined(__m68k__)
-    typedef unsigned short reg_t;
+    typedef int_fast8_t reg_t;
 #else
     typedef unsigned short reg_t;
 #endif
